+++ src/de/pseudonymisierung/mainzelliste/crypto/CryptoUtil.java
@@ -25,9 +25,14 @@
  */
 package de.pseudonymisierung.mainzelliste.crypto;
 
+import com.google.crypto.tink.CleartextKeysetHandle;
+import com.google.crypto.tink.JsonKeysetReader;
+import com.google.crypto.tink.KeysetHandle;
 import de.pseudonymisierung.mainzelliste.crypto.key.CryptoKey;
 import de.pseudonymisierung.mainzelliste.crypto.key.JCEKey;
 import de.pseudonymisierung.mainzelliste.crypto.key.KeyType;
+import de.pseudonymisierung.mainzelliste.crypto.key.TinkKeySet;
+import java.io.IOException;
 import java.security.GeneralSecurityException;
 import java.security.KeyFactory;
 import java.security.NoSuchAlgorithmException;
@@ -48,23 +53,35 @@ public class CryptoUtil {
         case RSA:
           PublicKey publicKey = wrappedKey.getKey(PublicKey.class);
           return new JCEAsymmetricEncryption(publicKey);
+        case TINK_HYBRID:
+          KeysetHandle keysetHandle = wrappedKey.getKey(KeysetHandle.class);
+          return new TinkHybridEncryption(keysetHandle);
       }
     } catch (IllegalArgumentException e) {
       throw new InvalidKeySpecException("The given crypto key '" +
           wrappedKey.getKey().getClass().getSimpleName() + "' and the encryption type '" +
           encryptionType + "' are incompatible ");
     } catch (GeneralSecurityException e) {
-      throw new UnsupportedOperationException("can't create " +  encryptionType + " encryption instance", e);
+      throw new UnsupportedOperationException(
+          "can't create " + encryptionType + " encryption instance", e);
     }
-    throw new IllegalArgumentException("the given encryption type " + encryptionType + " not supported yet");
+    throw new IllegalArgumentException(
+        "the given encryption type " + encryptionType + " not supported yet");
   }
 
   public static CryptoKey readKey(String keyType, byte[] encodedKey) {
     if (KeyType.valueOf(keyType.trim()) == KeyType.RSA_PUBLIC) {
       try {
-        PublicKey key = KeyFactory.getInstance("RSA").generatePublic(new X509EncodedKeySpec(encodedKey));
+        PublicKey key = KeyFactory.getInstance("RSA")
+            .generatePublic(new X509EncodedKeySpec(encodedKey));
         return new JCEKey(key);
       } catch (InvalidKeySpecException | NoSuchAlgorithmException e) {
+        throw new UnsupportedOperationException(e);
+      }
+    } else if (KeyType.valueOf(keyType.trim()) == KeyType.TINK_KEYSET) {
+      try {
+        return new TinkKeySet(CleartextKeysetHandle.read(JsonKeysetReader.withBytes(encodedKey)));
+      } catch (GeneralSecurityException | IOException e) {
         throw new UnsupportedOperationException(e);
       }
     }